chorale_core/views.py
from django.contrib.auth.models import User

from rest_framework import viewsets, decorators
from rest_framework.response import Response

from chorale_core.models import UserSerializer, Comment, CommentSerializer


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer

    @decorators.list_route()
    def me(self, request):
        serializer = self.serializer_class(request.user)
        return Response(serializer.data)


class CommentViewSet(viewsets.ModelViewSet):
    queryset = Comment.objects.all()
    serializer_class = CommentSerializer
    